pyproject.toml
[tool.poetry]
name = "ice-breaker"
version = "0.1.0"
description = "A GenAI application to gather information about a person, and find an ice breaker."
authors = ["basileazh <basile.elazhari@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
langchain = "^0.1.4"
langchain_openai = "^0.0.5"
openai="^1.3.5"
python-dotenv = "^1.0.0"
pydantic = "^2.5.2"
google-search-results = "^2.4.2"
apify-client = "^1.6.0"
python-slugify = "^8.0.1"
click = "^8.1.7"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]
pylint = "^2.7"
black = "^23.11"
flake8 = "^3.8.4"
mypy = "^1.7.0"
pre-commit = "^3.5"
isort = "^5.7.0"
wheel = "^0.36.2"
twine = "^3.4.1"
attrs = "^20.3.0"
pyyaml = "^6.0.1"
aiofiles = "^0.7.0"
types-requests = "^2.31.0.10"
jupyter = "^1.0.0"
types-python-slugify = "^8.0.0.3"

[tool.poetry.group.doc]
optional = true

[tool.poetry.group.doc.dependencies]
sphinx = "^7.2.6"
#sphinx-rtd-theme = "^0.5.2"
#sphinx-autodoc-typehints = "^1.11.1"
#sphinxcontrib-asyncio = "^0.3.0"

[tool.poetry.group.test]
optional = true

[tool.poetry.group.test.dependencies]
pytest = "^6.2"
pytest-cov = "^2.11"
pytest-mock = "^3.6"
pytest-pylint = "^0.18"
pytest-xdist = "^2.2"
pytest-asyncio = "^0.14"
pytest-mypy = "^0.8.0"
pytest-flake8 = "^1.0.7"
pytest-black = "^0.3.10"
pytest-isort = "^1.0.0"
pytest-randomly = "^3.5.0"
pytest-sugar = "^0.9.4"
pytest-timeout = "^1.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ice-breaker = "ice_breaker.scripts.ice_breaker:ice_breaker"

[tool.bandit]
exclude_dirs = ["tests", "docs", "build", "dist", ".venv", ".git"]
